pyproject.toml
[project]
name = "checklist-generator"
version = "0.1.0"
description = "A Python project to generate checklists from externally retrieved findings"
authors = [
    {name = "desp4c1to",email = "yyuannan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python (>=9.2.0,<10.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
